move callbacks, remove add_additional in predictions and args checks, improve seed and logging functionality
